requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
Flask==2.3.2
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pipreqs==0.4.13
requests==2.31.0
urllib3==2.0.3
Werkzeug==2.3.6
WTForms==3.0.1
yarg==0.1.9
zipp==3.15.0
email_validator>=1.1.2,<2.0.0
Flask-Bootstrap4==4.0.2